requirements.txt
streamlit
faiss-cpu
sentence-transformers
llama-cpp-python
